refactor(Home Page): Moves home page to pages folder.
