+++ ui/src/pages/HomePage/Register/index.js
