+++ ui/src/pages/HomePage/Register/styles.css
@@ -2,7 +2,7 @@
 .container {
   color: white;
   background-color: #f5aa34;
-  background-image: url("../../assets/geometric-pattern.png");
+  background-image: url("../../../assets/geometric-pattern.png");
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
